Add ACME Service Discovery plugin
Add a plugin that implements the ACME Service Discovery
specification[1].  This plugin is activated via by the `--discovery`
option.  There are two modes of operation.

- 'try' mode (the default) attempts to perform service discovery,
  but if a server is not found the default is used.

- 'force' mode does not fall back to the default server; if service
  discovery fails, the program dies

Parent domains for service discovery are taken from the
resolv.conf(5) search domain configuration (via the dnspython
'Resolver' class).  There are more possibilities for parent domain
selection (as discussed in the specification) but for this initial
implementation we will keep it simple.

Currently there is no log output to explain what happens if service
discovery failed.

[1] https://tools.ietf.org/html/draft-tweedale-acme-discovery-01